--- uikit/domql/NotificationIndicator/index.js
@@ -1,14 +1,0 @@
-'use strict'
-import { Flex } from '@symbo.ls/atoms'
-
-export const NotificationIndicator = {
-  extend: Flex,
-  props: {
-    text: '2',
-    fontSize: 'Z',
-    background: '#0474F2',
-    boxSize: 'fit-content fit-content',
-    padding: '- X+V',
-    round: 'Z'
-  }
-}
