+++ uikit/domql/index.js
@@ -7,7 +7,6 @@ export * from '@symbo.ls/box'
 export * from '@symbo.ls/indicatordot'
 export * from '@symbo.ls/check'
 export * from '@symbo.ls/titleparagraph'
-export * from '@symbo.ls/notificationindicator'
 export * from '@symbo.ls/field'
 export * from '@symbo.ls/search'
 export * from '@symbo.ls/uploadbutton'
