+++ uikit/domql/package.json
@@ -18,7 +18,6 @@
     "@symbo.ls/radiobutton": "latest",
     "@symbo.ls/toggleswitch": "latest",
     "@symbo.ls/titleparagraph": "latest",
-    "@symbo.ls/notificationindicator": "latest",
     "@symbo.ls/field": "latest",
     "@symbo.ls/search": "latest",
     "@symbo.ls/uploadbutton": "latest",
